Fix Safari back button
On Safari the bfcache caches the last state of the page and doesn't call the onload event when pressing the back button.
The problem is that we rely on the onload event to fade the page back in after we faded it out.

The solution is to add the js-has-loaded class back to the body on the last tick before opening the new page, otherwise the new page won't have it and nothing will show.